MedicalAssistant/requirements.txt
streamlit
pillow
PyMuPDF
langchain
langchain-community
langchain-groq
python-dotenv